requirements.txt
youtube-transcript-api
streamlit
google-generativeai
python-dotenv

